+++ master/singletons/global.gd
@@ -1,0 +1,6 @@
+extends Node
+# A singleton used to store global variables and utility functions.
+
+
+# Game modes which can be chosen at game start.
+enum GameMode { DEBUG, NORMAL, ENDURANCE, SUDDEN_DEATH }